+++ lib/src/ui.dart
@@ -4,6 +4,7 @@ export 'ui/board_controller.dart';
 export 'ui/hand.dart';
 export 'ui/move_animation.dart';
 export 'ui/overlay/board_overlay.dart';
+export 'ui/overlay/input_overlay.dart';
 export 'ui/overlay/label_overlay.dart';
 export 'ui/overlay/square_number_overlay.dart';
 export 'ui/piece.dart';